Pavel/ simple read-more (#2054)
* some more

* bump version

* added info-strip to confirmation dialog

* showErrorAlert

* showSuccessAlert

* positive button, more Alert methods

* little things

* mutation observables, housekeeping

* fix MetaTooltipDirective

* skipLibCheck

* revert package lock

* image card ratio

* delete confirmation, datepicker fixes

* fix card-image, add error config method

* some more

* added info-strip to confirmation dialog

* showErrorAlert

* showSuccessAlert

* positive button, more Alert methods

* little things

* mutation observables, housekeeping

* fix MetaTooltipDirective

* skipLibCheck

* revert package lock

* image card ratio

* delete confirmation, datepicker fixes

* fix card-image, add error config method

* some more

* added info-strip to confirmation dialog

* showErrorAlert

* showSuccessAlert

* positive button, more Alert methods

* little things

* mutation observables, housekeeping

* fix MetaTooltipDirective

* skipLibCheck

* revert package lock

* image card ratio

* delete confirmation, datepicker fixes

* fix card-image, add error config method

* simple read more

* ss

* linkText

* fix TextProps interface

* link focus
