+++ projects/ui-framework/src/lib/layout/read-more/read-more.module.ts
@@ -1,13 +1,13 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { TranslateModule } from '@ngx-translate/core';
+
 import { ReadMoreComponent } from './read-more.component';
-import { ButtonsModule } from '../../buttons/buttons.module';
-import { EventManagerPlugins } from '../../services/utils/eventManager.plugins';
 
 @NgModule({
-  imports: [CommonModule, ButtonsModule],
+  imports: [CommonModule, TranslateModule],
   declarations: [ReadMoreComponent],
   exports: [ReadMoreComponent],
-  providers: [EventManagerPlugins[0]],
+  providers: [],
 })
 export class ReadMoreModule {}
