+++ projects/ui-framework/src/lib/services/html/html-helpers.interface.ts
@@ -3,7 +3,10 @@ export interface Styles {
 }
 
 export interface TextProps {
-  [property: string]: number;
+  fontSize: number;
+  lineHeightPx: number;
+  lineHeight: number;
+  [prop: string]: number;
 }
 
 export interface NotEmptyChildren {
