+++ projects/ui-framework/src/lib/style/component-mixins/link.mixin.scss
@@ -10,7 +10,8 @@
     }
   }
 
-  &:hover {
+  &:hover,
+  &:focus {
     text-decoration: underline;
   }
 }
